Webview UX fixes
removed use of assert statement. (see http://stackoverflow.com/questions/1957645/when-to-use-an-assertion-and-when-to-use-an-exception)

removed use of v21 qualifier for layout. (Toolbar elevation is 4dp when not specified on some platforms and 0dp on some other platforms. Specifying elevation makes it consistent on all platforms with no performance loss)

returned FABs to bottom right.
Applied hide on scroll to bottom behavior on FABs.
Added toolbar refresh action for convenience. (Refresh without scrolling to top)